Expand db tests
Getting closer to having all old tests converted over.

[ci skip]
